requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.1.1
Faker==37.4.0
supabase==2.16.0
psycopg2-binary==2.9.10
Werkzeug==3.1.3 